Revert "Create new module for rpy2 utilities"
This reverts commit f04b9d38f0225449026f86c0cbd2cdea0bf9ee82.
