+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = ucb_prefect_tools
-version = 5.3.0
+version = 5.0.0
 author = James Ashby
 author_email = james.ashby@colorado.edu
 description = Tasks and tools for implementing Prefect data flows
@@ -14,7 +14,7 @@ classifiers =
 
 [options]
 packages = find:
-python_requires = >=3.10
+python_requires = >=3.8
 install_requires =
     boto3
     gitpython
@@ -29,5 +29,4 @@ install_requires =
     pyodbc
     pysmb
     pytz
-    rpy2
     s3fs
